Change products to calculate game from product ID
The correct game is now calculated based on the starting digit of the
product ID. For example, if the ID starts with 0, the product is from
Warhammer: 40k.

This allows NavLinks to function correctly from the search page, as the
previous functionality derived game from params which aren't available
on this page.